209 lelu

+++ src/main/java/fi/vaylavirasto/sillari/model/RouteTransportModel.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 @Data
 public class RouteTransportModel {
-    private long id;
-    private long routeId;
+    private Integer id;
+    private Integer routeId;
     private RouteTransportStatusModel currentStatus;
     private List<RouteTransportStatusModel> statusHistory;
     private List<SupervisionModel> supervisions;
